requirements.txt
# Core
python-dotenv>=1.0.0
pyyaml>=6.0
aiohttp>=3.9.0
pydantic>=2.0

# Blockchain
web3>=6.15.0

# Data & Analytics
pandas>=2.0.0
numpy>=1.26.0

# Dashboard (optional)
streamlit>=1.30.0
plotly>=5.18.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
